requirements.txt
Django==2.0.7
django-cors-headers==2.3.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
jdcal==1.4
openpyxl==2.5.4
Pillow==5.2.0
pytz==2018.5
